Debugged

+++ __init__.py
@@ -6,7 +6,7 @@ defaultBaseNames = {"user_files", ".git", ".gitignore", ".svn"}
 
 def backupUserFiles(self, sid):
     """Move user file's folder to a folder called files_backup in the add-on folder"""
-    if not os.path.exists(p):
+    if not os.path.exists(self._userFilesBackupPath()):
         os.mkdir(self._userFilesBackupPath())
     for filename in getUserOption("base names", defaultBaseNames):
         p = os.path.join(self.addonsFolder(sid), filename)
@@ -21,7 +21,8 @@ def restoreUserFiles(self, sid):
         if os.path.exists(os.path.join(self._userFilesBackupPath(), filename)):
             os.rename(os.path.join(self._userFilesBackupPath(), filename), os.path.join(self.addonsFolder(sid), filename))
 
-    os.rmdir(self._userFilesBackupPath())
+    if os.path.exists(self._userFilesBackupPath()):
+        os.rmdir(self._userFilesBackupPath())
 
 AddonManager.backupUserFiles = backupUserFiles
 AddonManager.restoreUserFiles = restoreUserFiles
